pyproject.toml
[build-system]
requires = [
    "setuptools",
    # 4110 only:
    # "cython",
    # "numpy==1.21.*",
]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
requires-python = ">=3.7"
license = {text = "MIT License"}
name = "instapy"
description = "Add cool filters to your pictures!"
readme = "README.md"
dependencies = [
    "instapy",
    "numpy",
    "pillow",
    "numba"
]


[project.scripts]
instapy = "instapy.cli:main"